requirements.txt
telebot
schedule
requests
python-dotenv
selenium 4.16.0